% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binom_trans}
\alias{binom_trans}
\title{Transformation of V matrix. From Multinomial regression to Logistic regression.}
\usage{
binom_trans(beta, beta_max = 10)
}
\arguments{
\item{beta}{A vector from logistic regression}

\item{beta_max}{Maximal value of beta, equivalent to L_inf penalty.}
}
\value{
A coefficient matrix used for logistic regression.
}
\description{
Transformation of V matrix. From Multinomial regression to Logistic regression.
}
