% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEGMoE_cv.R
\name{calcCriterion}
\alias{calcCriterion}
\title{Evaluate fitted NEGMoE obejct}
\usage{
calcCriterion(NEGMoE_obj, crit_list = c("AIC", "BIC"), ...)
}
\arguments{
\item{crit_list}{a list of evaluation metric.}

\item{...}{other parameters can be put into calcCriterion.}

\item{NEGMoE}{a NEGMoE of object with fitted parameters.}
}
\value{
A vector of evaluation result.
}
\description{
This function calculate evaluation of NEGMoE object.
}
\seealso{
\code{\link{createCVList}}
}
