% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcdf}
\alias{calcdf}
\title{Calculate degree of freedom}
\usage{
calcdf(NEGMoE_obj, output = "all")
}
\arguments{
\item{output}{whether degree of freedom with combined
components or compute within each component.}

\item{NEMoE_obj}{A NEGMoE object with fitted result.}
}
\value{
a list of degree of freedom in gating network and experts network.
}
\description{
This function calculate degree of freedom of NEGMoE object.
}
