% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_r_i_l}
\alias{check_r_i_l}
\title{Function to check if r_i is suitable for input of l layer}
\usage{
check_r_i_l(r_i, l)
}
\arguments{
\item{r_i}{An array for checking if contains any NA values.}

\item{l}{Number of layers}
}
\value{
An array of (d1, d2, l)
}
\description{
Function to check if r_i is suitable for input of l layer
}
