% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLasso.R
\name{chol_solve}
\alias{chol_solve}
\title{Inverse matrix solver use Cholesky decomposition}
\usage{
chol_solve(S)
}
\arguments{
\item{S}{A semi-definite matrix.}
}
\value{
S^{-1}
}
\description{
Inverse matrix solver use Cholesky decomposition
}
