% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{class_check}
\alias{class_check}
\title{Check whether the class of a vector or one hot matrix satisfies a valid class
i.e. the number elements within each class larger than min_obs}
\usage{
class_check(y, min_obs = 7)
}
\arguments{
\item{y}{Vector or matrix to check whether the observations of each classes is larger than min_obs}

\item{min_obs}{Minimal observations for each classes. By default is 7}
}
\value{
A logic value indicate whether all classes in y larger than min_obs.
}
\description{
Check whether the class of a vector or one hot matrix satisfies a valid class
i.e. the number elements within each class larger than min_obs
}
