% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clipping}
\alias{clipping}
\title{Clipping function to make the value within a range}
\usage{
clipping(x, max_x, min_x)
}
\arguments{
\item{x}{A vector to be clipping}

\item{max_x}{maximum value of the vector}

\item{min_x}{minimum value of the vector}
}
\value{
A clipped vector within the range
}
\description{
Clipping function to make the value within a range
}
