% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{createParameterList}
\alias{createParameterList}
\title{Create list of parameters for fitting NEMoE}
\usage{
createParameterList(
  lambda1 = 0.2,
  lambda2 = 0.03,
  alpha1 = 1,
  alpha2 = 1,
  stop = "all",
  gamma_func = inv_func,
  beta_max = 10,
  solver = "glmnet",
  stop_eps1 = 0.01,
  stop_eps2 = 0.01,
  itmax = 100,
  itmin = 3,
  itmax0 = 10,
  itmax1 = 100,
  itmax2 = 100,
  EM_opt = "EM",
  backtracking = "comp",
  adapt = T,
  init = "kmeans",
  ratio_max = 20,
  method = "icov",
  thresh_cor = 0.3,
  thresh_eig = 0.1,
  verbose = TRUE
)
}
\arguments{
\item{lambda1}{Penalty regularizer for the experts.}

\item{lambda2}{Penalty regularizer for the gating network.}

\item{alpha1}{Elastic net penalty value for experts.}

\item{alpha2}{Elastic net penalty value for gating network.}

\item{beta_max}{Maximal of coefficient. By default is 10.}

\item{itmax}{Maximium number of iteration in fitting NEMoE. By default is 100.}

\item{itmin}{Minimium number of iteration in fitting NEMoE. By default is 3.}

\item{backtracking}{Whether use backtracking during the iteration.
Can be chosen from}

\item{adapt}{whether use adaptive mode in optimization. By default is TRUE.}

\item{init}{Method for initialization.
Can be chosen from "rand", "kmeans" and "glmnet".
If init="rand" will use a dirichlet distribution initialing the latent class.
If init="kmeans" the latent class will initialized using kmeans
clustering of input for gating network.
If init="glmnet" will use a equal probability
with lasso as its corresponding coefficients in experts network.}

\item{verbose}{A logical input indicating whether the intermediate
steps will be printed.}

\item{stop_all}{Method of stop criterion. If stop_all = TRUE means that either
coefficient or loss function converge. If stop_all = FALSE means that both
coefficient and loss function converge.}

\item{EM_alg}{Method for Expecation maximization update.
Can be chosen from "EM", "CEM", "GEM", "SEM", "SAEM", "GEM".
By default is "EM".}

\item{early_stop}{A logical input indicate whether early stop when one of
the fitted latent class have zero variables selected (to save time).
By default is TRUE.}
}
\value{
A list contain parameters in fitting NEMoE.
}
\description{
This function create parameters that put into fitNEMoE function.
}
\seealso{
\code{\link{NEMoE_buildFromList}},
\code{\link{NEMoE_buildFromPhyloseq}}, \code{\link{fitNEMoE}}
}
