% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEGMoE_cv.R
\name{cvNEGMoE}
\alias{cvNEGMoE}
\title{Parameters tunning in NEGMoE}
\usage{
cvNEGMoE(NEGMoE_obj, verbose = T, ...)
}
\arguments{
\item{NEGMoE_obj}{a NEGMoE of object without cross validated parameters.}

\item{verbose}{A logical input indicating whether the intermediate
steps will be printed.}

\item{...}{Other parameters that can be passed to cvNEGMoE.}
}
\value{
A NEGMoE object with different lambdas.
}
\description{
This function calculate evaluation of NEGMoE object.
}
\seealso{
\code{\link{createCVList}}
}
