% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{dat_gen}
\alias{dat_gen}
\title{Generate simulated data}
\usage{
dat_gen(
  n,
  p = 20,
  q = 20,
  s = 5,
  K = 2,
  niu_beta = 2,
  r = 0.1,
  inv_signal = 0.8,
  inv_conn = 10,
  version = "probit"
)
}
\arguments{
\item{n}{Number of samples}

\item{p}{Number of variables of data matrix X (input of experts network)}

\item{q}{Number of variables of data matrix Z (input of gating network)}

\item{s}{Sparsity parameters, control non-zero coefficients in gating network coefficients.}

\item{K}{Number of mixture components}

\item{niu_beta}{Strength of beta for gating network}

\item{r}{Separation parameters for latent class. Larger r encourage more separation in latent classes.}

\item{inv_conn}{Number of non-zero components in inverse covariance matrix.}

\item{version}{link function, could be "logit" or "probit"}
}
\value{
A list contain simulated data, including data matrix, X, input of gating network Z, coefficients of
gating network beta_V, real mean of X mu, and real inverse covariance matrix of X Theta.
}
\description{
Generate simulated data
}
