% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.calcdf}
\alias{.calcdf}
\title{Calculate how many variables included in the matrix
If version = "all", return how many non-zero rows
Else return how many non-zero elements}
\usage{
.calcdf(X, type = "icov", version = "all", eps = 1e-08, intercept = TRUE)
}
\description{
Calculate how many variables included in the matrix
If version = "all", return how many non-zero rows
Else return how many non-zero elements
}
