% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_Elnet.R
\name{.Elnet_glmnet}
\alias{.Elnet_glmnet}
\title{Wrapper function for elastic net regression with glmnet.}
\usage{
.Elnet_glmnet(
  X,
  y,
  lambda,
  alpha = 1,
  R = NULL,
  beta_init = NULL,
  itmax = 1000,
  eps = 0.001,
  eps_var = 1e-05,
  intercept = TRUE
)
}
\description{
Wrapper function for elastic net regression with glmnet.
}
