% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likeli_ps.R
\name{.Likeli_binomial0}
\alias{.Likeli_binomial0}
\title{Calculate likelihood for binomial distribution}
\usage{
.Likeli_binomial0(y, prob, thresh = 1e-09, log = FALSE)
}
\description{
Calculate likelihood for binomial distribution
}
