% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likeli_ps.R
\name{.Likeli_multi}
\alias{.Likeli_multi}
\title{Calculate likelihood function of multinomial distribution with regression input.}
\usage{
.Likeli_multi(X, B, y, thresh = 1e-09, log = FALSE)
}
\description{
Calculate likelihood function of multinomial distribution with regression input.
}
