% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likeli_ps.R
\name{.Likeli_obs}
\alias{.Likeli_obs}
\title{Calculate the observed likelihood function, i.e. likelihood function for mixture distribution.}
\usage{
.Likeli_obs(X, Z, y, beta_V, beta_W, log = FALSE, thresh = 1e-06)
}
\description{
Calculate the observed likelihood function, i.e. likelihood function for mixture distribution.
}
