% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likeli_ps.R
\name{.LL_all}
\alias{.LL_all}
\title{Function to calculate loglikelihood function including observed loglikelihood,
penalized loglikehood, complete loglikelihood and penalized complete loglikelihood.}
\usage{
.LL_all(
  X,
  Z,
  y,
  beta_V,
  beta_W,
  lambda1,
  lambda2,
  alpha1,
  alpha2,
  multiply = F
)
}
\description{
Function to calculate loglikelihood function including observed loglikelihood,
penalized loglikehood, complete loglikelihood and penalized complete loglikelihood.
}
