% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likeli_ps.R
\name{.LL_all_mat}
\alias{.LL_all_mat}
\title{Function to calculate loglikelihood function with list input.}
\usage{
.LL_all_mat(
  X_list,
  Z,
  y,
  beta_V,
  beta_W,
  lambda1_mat,
  lambda2,
  alpha1_mat,
  alpha2,
  multiply = F
)
}
\description{
Function to calculate loglikelihood function with list input.
}
