% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likeli_ps.R
\name{.LL_all_mvr}
\alias{.LL_all_mvr}
\title{Wrapper function to calculate all log-likelihood function for multivariate elastic net regression}
\usage{
.LL_all_mvr(
  X,
  Z,
  r_i,
  Y = NULL,
  V,
  W,
  lambda1,
  lambda2,
  alpha1,
  alpha2,
  multiply = T,
  mode = "ps",
  ratio_max = 20
)
}
\description{
Wrapper function to calculate all log-likelihood function for multivariate elastic net regression
}
