% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLasso.R
\name{.LLglasso}
\alias{.LLglasso}
\title{Calculate Likelihood function for gglasso}
\usage{
.LLglasso(X, mu = NULL, Theta, weight = NULL, reduce = T, type = "icov")
}
\description{
Calculate Likelihood function for gglasso
}
