% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likeli_ps.R
\name{.LLikeli_mvr}
\alias{.LLikeli_mvr}
\title{Likelihood function for multivariate normal distribution (LL(Y-XB))}
\usage{
.LLikeli_mvr(X, B, Y = NULL, sgmY = NULL, R = NULL, intercept = T, reduce = T)
}
\description{
Likelihood function for multivariate normal distribution (LL(Y-XB))
}
