% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likeli_ps.R
\name{.LogLikeli_complete}
\alias{.LogLikeli_complete}
\title{Calculate complete log likelihood function, i.e. the corresponding Q function.}
\usage{
.LogLikeli_complete(X, Z, y, beta_V, beta_W, reduce = FALSE)
}
\description{
Calculate complete log likelihood function, i.e. the corresponding Q function.
}
