% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likeli_ps.R
\name{.obj_elnet}
\alias{.obj_elnet}
\title{objective function for elastic net regression}
\usage{
.obj_elnet(y, X, R = NULL, beta, lambda, alpha, intercept = F, multiply = F)
}
\description{
objective function for elastic net regression
}
