% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likeli_ps.R
\name{.obj_logistic}
\alias{.obj_logistic}
\title{objective function for logistic regression}
\usage{
.obj_logistic(
  X,
  beta,
  y,
  R = NULL,
  lambda,
  alpha,
  intercept = F,
  multiply = F,
  residual_k = 1
)
}
\description{
objective function for logistic regression
}
