% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likeli_ps.R
\name{.obj_multi}
\alias{.obj_multi}
\title{objective function for sparse multinomial regression}
\usage{
.obj_multi(X, beta, y, R = NULL, lambda, alpha, intercept = F, multiply = F)
}
\description{
objective function for sparse multinomial regression
}
