% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likeli_ps.R
\name{.PLL}
\alias{.PLL}
\title{Calculate penalized likelihood function.}
\usage{
.PLL(
  X,
  Z,
  y,
  beta_V,
  beta_W,
  version = "complete",
  lambda1 = NULL,
  lambda2 = NULL,
  alpha1 = 1,
  alpha2 = 1,
  multiply = F
)
}
\description{
Calculate penalized likelihood function.
}
