% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likeli_ps.R
\name{.PLL_mvr}
\alias{.PLL_mvr}
\title{Penalized likelihood function for multivariate normal distribution (PLL(Y-XB))}
\usage{
.PLL_mvr(
  X,
  B,
  Y = NULL,
  sgmY = NULL,
  R = NULL,
  lambda,
  alpha = 1,
  intercept = T,
  multiply = T,
  div_sgm = T
)
}
\description{
Penalized likelihood function for multivariate normal distribution (PLL(Y-XB))
}
