% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLasso.R
\name{.PLLglasso}
\alias{.PLLglasso}
\title{Calculate penalized log-likelihood for graphical lasso.}
\usage{
.PLLglasso(X, mu = NULL, Theta, lambda, weight = NULL, type = "icov")
}
\description{
Calculate penalized log-likelihood for graphical lasso.
}
