% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likeli_ps.R
\name{.Q_func}
\alias{.Q_func}
\title{Q function for given current parameters.}
\usage{
.Q_func(
  X,
  Z,
  y,
  beta_V,
  beta_W,
  beta_V0,
  beta_W0,
  lambda1 = NULL,
  lambda2 = NULL,
  alpha1 = 1,
  alpha2 = 1,
  reduced = FALSE,
  multiply = F
)
}
\description{
Q function for given current parameters.
}
