% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likeli_ps.R
\name{.r_i_arr}
\alias{.r_i_arr}
\title{Calculate the probability for each latent class with multiple input.}
\usage{
.r_i_arr(X_list, Z, y, pi_old, beta_W, beta_V, r_i_eps = 1e-06)
}
\description{
Calculate the probability for each latent class with multiple input.
}
