% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_multi.R
\name{.sMulti_single}
\alias{.sMulti_single}
\title{Inner loop (Elastic net) for sparse multinomial regression with proximal Newtown updating.}
\usage{
.sMulti_single(
  X,
  y,
  beta0,
  residual_k,
  R,
  backtracking = T,
  lambda,
  alpha,
  itmax = 1000,
  beta_max = 10,
  eps = 1e-05,
  version = "my",
  intercept = TRUE
)
}
\description{
Inner loop (Elastic net) for sparse multinomial regression with proximal Newtown updating.
}
