% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_multi.R
\name{.sMulti_step}
\alias{.sMulti_step}
\title{Inner loop (Coordinate descent) for sparse multinomial regression with proximal Newtown updating.}
\usage{
.sMulti_step(
  X,
  y,
  beta0,
  lambda,
  R = NULL,
  backtracking = T,
  alpha,
  itmax = 1000,
  beta_max = 10,
  version = version
)
}
\description{
Inner loop (Coordinate descent) for sparse multinomial regression with proximal Newtown updating.
}
