% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_Elnets.R
\name{Elnet_multi}
\alias{Elnet_multi}
\title{A wrapper function for Elastic net for multivariate response (Y multivariate)}
\usage{
Elnet_multi(
  X,
  Y,
  lambda,
  alpha = 1,
  R = NULL,
  beta_init = NULL,
  itmax = 1000,
  eps = 0.001,
  eps_var = 1e-05,
  intercept = TRUE,
  version = "glmnet"
)
}
\arguments{
\item{X}{covariate for elastic net regression (n x p).}

\item{Y}{response variable for elastic net regression (n x q).}

\item{lambda}{parameter lambda for elastic net.}

\item{alpha}{parameter lambda for elastic net(if alpha = 1 for lasso).}

\item{R}{weight matrix for each observations (n x 1).}

\item{beta_init}{initial value for beta.}

\item{itmax}{maximum number of iterations. By default is 1e3.}

\item{eps}{A numeric threshold for coordinate descent. By default is 1e-3.}

\item{eps_var}{A numeric threshold to exclude variables with variance smaller than the threshold.
By default is 1e-5.}

\item{intercept}{A logic variable to indicate whether include intercept terms.}

\item{version}{If version = "glmnet" use glmnet package for fitting or use
self-written coordinate descent.}
}
\value{
a matrix of regression coefficients.
}
\description{
A wrapper function for Elastic net for multivariate response (Y multivariate)
}
