% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elnet_pseudo.R
\name{Elnet_ps}
\alias{Elnet_ps}
\title{A wrapper function for Elastic net for each variable (pseudo-likelihood)}
\usage{
Elnet_ps(
  X,
  lambda,
  alpha = 1,
  R = NULL,
  beta_init = NULL,
  itmax = 1000,
  eps = 0.001,
  eps_var = 1e-05,
  intercept = TRUE,
  version = "glmnet"
)
}
\arguments{
\item{X}{data matrix for calculate pseudo-likelihood using elastic net (n x p).}

\item{lambda}{parameter lambda for elastic net.}

\item{alpha}{parameter lambda for elastic net(if alpha = 1 for lasso).}

\item{R}{weight matrix for each observations (n x 1).}

\item{beta_init}{initial value for beta.}

\item{itmax}{maximum number of iterations.}

\item{eps}{A numeric threshold for coordinate descent.}

\item{eps_var}{A numeric threshold to exclude variables with variance smaller than the threshold}

\item{intercept}{A logic variable to indicate whether include intercept terms.}

\item{version}{If version = "glmnet" use glmnet package for fitting or use
self-written coordinate descent.}
}
\value{
a matrix of regression coefficients.
}
\description{
A wrapper function for Elastic net for each variable (pseudo-likelihood)
}
