% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ext_param}
\alias{ext_param}
\title{Extenion of parameters. To make the number of penalty the same number of components}
\usage{
ext_param(lambda, K)
}
\arguments{
\item{lambda}{Penalty parameters.}

\item{K}{Number of components.}
}
\value{
A matrix suits for NEGMoE.
}
\description{
Extenion of parameters. To make the number of penalty the same number of components
}
