% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ext_param_mat}
\alias{ext_param_mat}
\title{Matrix version of Extension of parameters.}
\usage{
ext_param_mat(lambda, K, l)
}
\arguments{
\item{lambda}{Penalty parameters.}

\item{K}{Number of components.}

\item{l}{Number of layers}
}
\value{
A list suits for NEGMoE.
}
\description{
Matrix version of Extension of parameters.
}
