% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{getCoef}
\alias{getCoef}
\title{Get coefficients of fitted NEGMoE object}
\usage{
getCoef(NEGMoE_obj)
}
\arguments{
\item{NEMoE_obj}{A NEGMoE object.}
}
\value{
A list contain fitted result.
}
\description{
This function get coefficients in NEGMoE object.
}
\seealso{
\code{\link{fitNEMoE}}
}
