% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{getLL}
\alias{getLL}
\title{Get Likelihood of fitted NEMoE object}
\usage{
getLL(NEGMoE_obj)
}
\arguments{
\item{NEMoE_obj}{A NEMoE object.}
}
\value{
A dataframe contain fitted result.
}
\description{
This function get Likelihood result in NEMoE object.
}
\seealso{
\code{\link{fitNEMoE}}
}
