% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLasso.R
\name{GLasso}
\alias{GLasso}
\title{Graphical lasso algorithm}
\usage{
GLasso(
  X,
  mu = NULL,
  lambda,
  weight = NULL,
  itmax = 100,
  precise = T,
  Theta_init = NULL
)
}
\arguments{
\item{X}{A data matrix of n observations and p variables.}

\item{mu}{A vector of means for each variable(p x 1).}

\item{lambda}{penalty parameters for Glasso.}

\item{weight}{weighted vector for each observations (n x 1). If is NULL, will use all 1 vectors.}

\item{itmax}{Maximum number of iterations. By default is 1e2.}

\item{precise}{A logic variable indicate whether to use adaptive or ordinary graphical lasso.}

\item{Theta_init}{Initial value of Theta. If is NULL, will use a diagonal matrix.}
}
\value{
An estimation of inverse covariance matrix (Theta).
}
\description{
Graphical lasso algorithm
}
