% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hard_prob}
\alias{hard_prob}
\title{A function for converting a soft probability matrix to factor or one hot matrix.}
\usage{
hard_prob(prob, idx = FALSE)
}
\arguments{
\item{prob}{a matrix to convert.}

\item{idx}{the type of output. If idx= TRUE will return the index of largest probability of each row. If idx= FALSE will return a one-hot matrix.}
}
\value{
a index of largest probability or one hot matrix.
}
\description{
A function for converting a soft probability matrix to factor or one hot matrix.
}
