% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hard_thresh}
\alias{hard_thresh}
\title{hard threshold function}
\usage{
hard_thresh(X, thresh = 1e-08)
}
\arguments{
\item{X}{Input of hard threshold function}

\item{thresh}{Threshold}
}
\value{
Hard threshold of X
}
\description{
hard threshold function
}
