% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv_func}
\alias{inv_func}
\title{Inverse function. Using for SAEM optimization.}
\usage{
inv_func(it)
}
\arguments{
\item{it}{number of iterations}
}
\value{
\frac{1}{it}
}
\description{
Inverse function. Using for SAEM optimization.
}
