% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLasso.R
\name{LL_Glasso}
\alias{LL_Glasso}
\title{All kinds of likelihood function for graphical lasso.}
\usage{
LL_Glasso(X, Z, beta_V, Theta, mu, r_i, lambda1, lambda2, ratio_max = 20)
}
\arguments{
\item{X}{Input matrix of X for experts network.}

\item{Z}{Input matrix of Z for gating network.}

\item{beta_V}{Coefficients matrix for gating network.}

\item{Theta}{An array for estimated inverse covariance matrices.}

\item{mu}{mean vector for each experts.}

\item{r_i}{likelihood ratio for each components.}

\item{lambda1}{Penalties for graphical lasso.}

\item{lambda2}{Penalties for gating networks.}

\item{ratio_max}{clipping for the maximium of ratio to be in one cluster.}
}
\value{
A vector contain Log-likelihood, Penalized Log-likelihood, Complete log-likelihood and Penalized complete log-likelihood.
}
\description{
All kinds of likelihood function for graphical lasso.
}
