% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Criterion.R
\name{MoE_ICL}
\alias{MoE_ICL}
\title{Calculate ICL of a RMoE result}
\usage{
MoE_ICL(EM_res, eps = 1e-05)
}
\arguments{
\item{EM_res}{A list of RMoE result.}

\item{eps}{A threshold for minimal coefficient for calculate degree of freedom.}
}
\value{
A numerical variable of ICL statistics of the fitted RMoE model.
}
\description{
Calculate ICL of a RMoE result
}
