% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEGMoE_class.R
\docType{class}
\name{NEGMoE-class}
\alias{NEGMoE-class}
\alias{NEGMoE}
\title{NEGMoE class}
\description{
NEGMoE class
}
\section{Slots}{

\describe{
\item{\code{X}}{A matrix of microbiome data, rows are samples.}

\item{\code{Z}}{A matrix of nutrition intake.}

\item{\code{Y}}{A vector of response or NULL.}

\item{\code{type}}{Type of model, can be chosen from "icov", "pseudoLL" or "mElnet".
"mElnet" corresponding to multivariate elastic net model. (Y must given)
"icov" corresponding to sparse inverse covariance matrix estimation. (graphical lasso)
"pseudoLL" corresponding to pseudo likelihood method. (mb in huge package)
By default, will use "pseudoLL".}

\item{\code{params}}{A list of parameters for fitting NEGMoE.}

\item{\code{K}}{A number of fitting componenets in NEGMoE (>=2).}

\item{\code{cvParams}}{A list of parameters for selecting parameters of NEGMoE
using cross validation.}

\item{\code{cvResult}}{A dataframe of cross validation results of different parameters
of fitting NEGMoE.}

\item{\code{NEGMoE_output}}{A list NEGMoE fitting result.}

\item{\code{taxLevel}}{A character indicates selected taxa level.}

\item{\code{ResponseLevel}}{A character indicates levels of response variables.}

\item{\code{standardize}}{A Logical variable indicate whether standardize input data.}
}}

