% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{NEGMoE_buildFromList}
\alias{NEGMoE_buildFromList}
\title{Build NEGMoE object from list, dataframe or matrix}
\usage{
NEGMoE_buildFromList(
  X,
  Z,
  type = "pseudoLL",
  Y = NULL,
  K = NULL,
  cvParams = list(),
  taxLevel = NULL,
  taxTab = NULL,
  standardize = TRUE,
  ...
)
}
\arguments{
\item{X}{A list of input matrix (Microbiome)
data of different taxonomic level. If it is a dataframe or matrix, will
convert to a list automatically.}

\item{Z}{A dataframe of matrix of gating networks.}

\item{type}{Type of model, can be chosen from "icov", "pseudoLL" or "mElnet".
"mElnet" corresponding to multivariate elastic net model. (Y must given)
"icov" corresponding to sparse inverse covariance matrix estimation. (graphical lasso)
"pseudoLL" corresponding to pseudo likelihood method. (mb in huge package)
By default, will use "pseudoLL".}

\item{Y}{A multivariate outcome matrix. If is NULL, type will be automatically change to "pseudoLL".}

\item{K}{A integer of number of latent class.}

\item{cvParams}{A list of cross validation parameters.}

\item{taxLevel}{A character of selected name of taxonomic levels.}

\item{taxTab}{A dataframe of taxonomic table.}

\item{standardize}{Logical flag for x variable standardization.
Default is standardize=TRUE.}

\item{...}{Other parameters can pass to NEMoE_buildFromList.
See \code{\link{createParameterList}}}
}
\value{
A NEGMoE object.
}
\description{
This function build NEGMoE object from List, dataframe or matrix.
}
