% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEGMoE_crit.R
\name{NEGMoE_predict}
\alias{NEGMoE_predict}
\title{Make predictions from a fitted "NEGMoE" object.}
\usage{
NEGMoE_predict(
  NEGMoE_obj,
  X_new,
  Z_new,
  Y_new = NULL,
  out = "likelihood",
  full = TRUE
)
}
\arguments{
\item{NEGMoE_obj}{A NEMoE_obj object with fitted parameters.}

\item{X_new}{A matrix of new X.}

\item{Z_new}{A matrix of new Y.}

\item{Y_new}{A matrix of new Z.}

\item{full}{Whether output the result in both gating network
and experts network.}
}
\value{
A number of fitted (pseudo) Likelihood or predicted response.
}
\description{
This function predict new data using.
}
