% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{onehot}
\alias{onehot}
\title{One hot encoding to convert a factor result to one hot matrix}
\usage{
onehot(x)
}
\arguments{
\item{x}{A factor.}
}
\value{
One hot encoding of x.
}
\description{
One hot encoding to convert a factor result to one hot matrix
}
