% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEGMoE_plot.R
\name{plotExperts}
\alias{plotExperts}
\title{Plot experts network}
\usage{
plotExperts(NEGMoE_obj, thresh = 0.01, pseudo_type = "or")
}
\arguments{
\item{NEGMoE_obj}{a NEGMoE object with fitted output.}
}
\value{
A list of graph of fitted coefficients of each latent class
on each level.
}
\description{
This function plot the estimated coefficients of each latent class
on each level.
}
