% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEGMoE_plot.R
\name{plotGating}
\alias{plotGating}
\title{Plot gating network}
\usage{
plotGating(NEGMoE_obj, PCs = c(1, 2))
}
\arguments{
\item{NEGMoE_obj}{a NEGMoE object with fitted output.}

\item{PCs}{Visualization of selected Principal components.}
}
\value{
A graph of PCA plot of nutrition intake and estimated latent classes
and its corresponding loadings.
}
\description{
This function plot the PCA of fitted latent class
and their corresponding loadings.
}
