% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEGMoE_plot.R
\name{plotScatter}
\alias{plotScatter}
\title{Plot experts network}
\usage{
plotScatter(NEGMoE_obj, topdiff = 3, Varlist = NULL)
}
\arguments{
\item{NEGMoE_obj}{a NEGMoE object with fitted output.}
}
\value{
A list of graph of fitted coefficients of each latent class
on each level.
}
\description{
This function do scatter plot for differential correlations.
}
