% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likeli_ps.R
\name{prob_multi}
\alias{prob_multi}
\title{Calculate probability for multinomial regression}
\usage{
prob_multi(X, B)
}
\arguments{
\item{X}{the input data matrix (Need to add intercept term).}

\item{B}{coefficient of multinomial regression.
return the probability for multinomial regression.}
}
\description{
Calculate probability for multinomial regression
}
