% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLasso.R
\name{proj_sd}
\alias{proj_sd}
\title{Project a matrix to semi-definite space.}
\usage{
proj_sd(X, thresh = 0.1)
}
\arguments{
\item{X}{Input of the matrix}

\item{thresh}{threshold for eigenvalue of output semidefinite matrix.}
}
\value{
A semi-definite matrix by iterative projection.
}
\description{
Project a matrix to semi-definite space.
}
