% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rCate}
\alias{rCate}
\title{Categorical distribution}
\usage{
rCate(p)
}
\arguments{
\item{p}{probability for categorical distribution}
}
\value{
Random number follows Cat(p) distribution
}
\description{
Categorical distribution
}
