% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Generation data follows Dirichlet distribution.}
\usage{
rdirichlet(n, a)
}
\arguments{
\item{n}{Number of samples}

\item{a}{A vector of length p for Dirichlet distribution parameters.}
}
\value{
A matrix of n x p each row iid follows Dir(a)
}
\description{
Generation data follows Dirichlet distribution.
}
