% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdm}
\alias{rdm}
\title{Generation data follows Dirichlet-Multinomial distribution}
\usage{
rdm(n, p, lib_size, pi, over_disp = 0.3, type = "fixed")
}
\arguments{
\item{n}{Number of samples}

\item{p}{Number of variables.}

\item{lib_size}{library size for each row.(rowSums for each samples)}

\item{pi}{probability for multinomial distribution. If not provide, will generate by rnorm/sum(rnorm)}

\item{over_disp}{Overdispersion parameters. By default is 0.3.}

\item{type}{type of libaray size, if is "fixed" row sum will be the same or will follows poisson distribution.}
}
\value{
A matrix of n x p each row iid follows Dir(a)
}
\description{
Generation data follows Dirichlet-Multinomial distribution
}
