% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{setParam}
\alias{setParam}
\title{Set parameters of fitting NEMoE object}
\usage{
setParam(NEGMoE_obj, ...)
}
\arguments{
\item{...}{Other parameters can pass to setParam.}

\item{NEMoE_obj}{A NEMoE object.}
}
\value{
A NEMoE with user input parameters.
}
\description{
This function set parameters in NEMoE object.
}
\seealso{
\code{\link{createParameterList}}
}
