% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_multi.R
\name{sMulti_glmnet}
\alias{sMulti_glmnet}
\title{Wrapper function for sparse Multinomial regression}
\usage{
sMulti_glmnet(
  X,
  y,
  R = NULL,
  lambda,
  alpha = 1,
  beta_max = 10,
  intercept = TRUE,
  var_eps = 1e-05
)
}
\arguments{
\item{X}{description}

\item{y}{description}

\item{R}{description}

\item{lambda}{description}

\item{alpha}{description}

\item{beta_max}{description}

\item{intercept}{description}

\item{var_eps}{description}
}
\value{
description
}
\description{
Wrapper function for sparse Multinomial regression
}
