% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{V_trans}
\alias{V_trans}
\title{Transformation of V matrix. From Logistic regression to multinomial regression.}
\usage{
V_trans(V, beta_max = 10)
}
\arguments{
\item{V}{Regression coefficient of gating network}

\item{beta_max}{Maximal value of beta, equivalent to L_inf penalty.}
}
\value{
A coefficient matrix used for multinomial regression.
}
\description{
Transformation of V matrix. From Logistic regression to multinomial regression.
}
