% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{W_trans}
\alias{W_trans}
\title{Transformation of W list. From Logistic regression to Multinomial regression.}
\usage{
W_trans(W)
}
\arguments{
\item{W}{Regression coefficient of gating network}
}
\value{
A coefficient matrix from W list.
}
\description{
Transformation of W list. From Logistic regression to Multinomial regression.
}
